+++ packages/icons-react/documentation/IconsExample.tsx
@@ -25,6 +25,9 @@ import {
     WarningIcon,
     SuccessIcon,
     LinkIcon,
+    DragIcon,
+    GreenCheckIcon,
+    RedCrossIcon,
 } from "../src";
 import { IconVariant } from "../src/icons/types";
 import { IconExample } from "./internal/IconExample";
@@ -59,7 +62,10 @@ const allIcons = [
     ChevronUpIcon,
     CopyIcon,
     DotsIcon,
+    DragIcon,
+    GreenCheckIcon,
     QuestionIcon,
+    RedCrossIcon,
     InfoIcon,
     ErrorIcon,
     WarningIcon,