+++ packages/icons-react/src/icons/dots/DotsSmall.tsx
@@ -1,10 +1,10 @@
 import React, { FC, SVGAttributes } from "react";
 
 export const DotsSmall: FC<SVGAttributes<SVGElement>> = (props) => (
-    <svg viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 20 20" {...props}>
         <path
-            d="M6 10a1 1 0 11-2 0 1 1 0 012 0zM11 10a1 1 0 11-2 0 1 1 0 012 0zM16 10a1 1 0 11-2 0 1 1 0 012 0z"
             fill="currentColor"
+            d="M11.5 10a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0Zm5.5 0a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0ZM6 10a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0Z"
         />
     </svg>
 );