+++ packages/icons-react/src/icons/dots/DotsSmallBold.tsx
@@ -1,10 +1,10 @@
 import React, { FC, SVGAttributes } from "react";
 
 export const DotsSmallBold: FC<SVGAttributes<SVGElement>> = (props) => (
-    <svg viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 20 20" {...props}>
         <path
-            d="M6 10a2 2 0 11-4 0 2 2 0 014 0zM12 10a2 2 0 11-4 0 2 2 0 014 0zM18 10a2 2 0 11-4 0 2 2 0 014 0z"
             fill="currentColor"
+            d="M6 10a2 2 0 1 1-4 0 2 2 0 0 1 4 0Zm6 0a2 2 0 1 1-4 0 2 2 0 0 1 4 0Zm6 0a2 2 0 1 1-4 0 2 2 0 0 1 4 0Z"
         />
     </svg>
 );