+++ packages/icons-react/src/icons/drag/DragMediumBold.tsx
@@ -1,0 +1,10 @@
+import React, { FC, SVGAttributes } from "react";
+
+export const DragMediumBold: FC<SVGAttributes<SVGElement>> = (props) => (
+    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" {...props}>
+        <path
+            fill="currentColor"
+            d="M8 7a2 2 0 1 1 0-4 2 2 0 0 1 0 4Zm8 0a2 2 0 1 1 0-4 2 2 0 0 1 0 4Zm-8 7a2 2 0 1 1 0-4 2 2 0 0 1 0 4Zm8 0a2 2 0 1 1 0-4 2 2 0 0 1 0 4Zm-8 7a2 2 0 1 1 0-4 2 2 0 0 1 0 4Zm8 0a2 2 0 1 1 0-4 2 2 0 0 1 0 4Z"
+        />
+    </svg>
+);